refactor: move css to styles directory, remove uncommon constants from common contants file, add common styles to index.css
