+++ src/styles/PopMessage.module.css
@@ -1,5 +1,5 @@
 :root {
-  --size: 1rem;
+  --pop-size: 1rem;
   --notification-duration: 2s;
 }
 
@@ -12,8 +12,8 @@
   left:calc(50% - 180px / 2);
   box-sizing:border-box;
   border-radius:15px;
-  font-size: var(--size);
-  line-height: var(--size);
+  font-size: var(--pop-size);
+  line-height: var(--pop-size);
 }
 .animateOpen {
   animation-iteration-count: 1;
@@ -28,19 +28,14 @@
     transform: translateY(0);
   }
   20% {
-    height: calc(var(--size) * 3);
+    height: calc(var(--pop-size) * 3);
     padding: 0.5rem;
-    transform: translateY( calc(var(--size)));
-  }
-  50% {
-    height: calc(var(--size) * 3);
-    padding: 0.5rem;
-    transform: translateY(calc(var(--size)));
+    transform: translateY(calc(var(--pop-size)));
   }
   80% {
-    height: calc(var(--size) * 3);
+    height: calc(var(--pop-size) * 3);
     padding: 0.5rem;
-    transform: translateY( calc(var(--size)));
+    transform: translateY(calc(var(--pop-size)));
   }
   100% {
     height: 0;
